Break out Modules out of Features
