+++ projects/plugins/boost/.gitattributes
@@ -3,7 +3,7 @@
 # /src/js/example.min.js  production-include
 
 app/assets/dist/**                              production-include
-app/features/image-guide/dist/**                production-include
+app/modules/image-guide/dist/**                production-include
 jetpack_vendor/**                               production-include
 vendor/autoload.php                             production-include
 vendor/autoload_functions.php                   production-include
