+++ projects/plugins/boost/app/features/speed-score/Speed_Score.php
@@ -9,7 +9,7 @@
 
 namespace Automattic\Jetpack_Boost\Features\Speed_Score;
 
-use Automattic\Jetpack_Boost\Features\Optimizations\Features;
+use Automattic\Jetpack_Boost\Modules\Features;
 use Automattic\Jetpack_Boost\Lib\Utils;
 
 /**
