+++ projects/plugins/boost/app/modules/image-guide/src/_unused-image-observer.ts
