+++ projects/plugins/boost/app/modules/image-guide/src/analytics.ts
