+++ projects/plugins/boost/app/modules/image-guide/src/initialize.ts
