+++ projects/plugins/boost/app/modules/image-guide/src/stores/GuideState.ts
