+++ projects/plugins/boost/app/modules/image-guide/src/stores/MeasurableImageStore.ts
