+++ projects/plugins/boost/app/modules/image-guide/src/types.ts
