+++ projects/plugins/boost/app/modules/image-guide/src/ui/AdminBarToggle.svelte
