+++ projects/plugins/boost/app/modules/image-guide/src/ui/Bubble.svelte
