+++ projects/plugins/boost/app/modules/image-guide/src/ui/Main.svelte
