+++ projects/plugins/boost/app/modules/image-guide/src/ui/Portal.svelte
