+++ projects/plugins/boost/app/modules/image-guide/src/ui/assets/Checkmark.svelte
