+++ projects/plugins/boost/app/modules/image-size-report/Image_Size_Report.php
