+++ projects/plugins/boost/app/modules/optimizations/lazy-images/class-lazy-images.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Automattic\Jetpack_Boost\Features\Optimizations\Lazy_Images;
+namespace Automattic\Jetpack_Boost\Modules\Optimizations\Lazy_Images;
 
 use Automattic\Jetpack\Jetpack_Lazy_Images;
 use Automattic\Jetpack_Boost\Contracts\Feature;
