+++ projects/plugins/boost/app/modules/optimizations/minify/functions-helpers.php
@@ -1,7 +1,7 @@
 <?php
 
-use Automattic\Jetpack_Boost\Features\Optimizations\Minify\Config;
-use Automattic\Jetpack_Boost\Features\Optimizations\Minify\Dependency_Path_Mapping;
+use Automattic\Jetpack_Boost\Modules\Optimizations\Minify\Config;
+use Automattic\Jetpack_Boost\Modules\Optimizations\Minify\Dependency_Path_Mapping;
 
 // @todo - refactor this. Dump of functions from page optimize.
