+++ projects/plugins/boost/app/rest-api/endpoints/Feature_Status.php
@@ -22,15 +22,15 @@ class Feature_Status implements Endpoint {
 			);
 		}
 		$module_slug = $request['slug'];
-		$state       = new Status( $module_slug );
+		$status       = new Status( $module_slug );
 
 		// @TODO: Validate that the module exists?
 
 		$new_status = (bool) $params['status'];
-		$success    = $state->update( (bool) $params['status'] );
+		$success    = $status->update( (bool) $params['status'] );
 
 		return rest_ensure_response(
-			$success ? $new_status : $state->is_enabled()
+			$success ? $new_status : $status->is_enabled()
 		);
 	}
