+++ projects/plugins/boost/tsconfig.json
@@ -1,6 +1,8 @@
 {
 	"extends": "jetpack-js-tools/tsconfig.base.json",
-	"include": [ "app/assets/src/js/**/*", "app/features/image-guide/src/**/*" ],
+	"include": [ "app/assets/src/js/**/*",
+      "app/modules/image-guide/src/**/*"
+    ],
 	"exclude": [ "node_modules/*", "app/assets/dist/*", "tests/e2e/*" ],
 	"compilerOptions": {
 		"target": "es6",
